feat: fixed duplicate checkin bug, added confirmation modal

+++ src/app/components/menu/menu.component.ts
@@ -1,10 +1,12 @@
-import { DialogRef } from '@angular/cdk/dialog';
+import { Dialog, DialogRef } from '@angular/cdk/dialog';
 import { Component, inject, signal } from '@angular/core';
-import { TranslatePipe } from '@ngx-translate/core';
+import { TranslatePipe, TranslateService } from '@ngx-translate/core';
 import { SocialLogin } from '@capgo/capacitor-social-login';
 import { SocketAddress } from 'node:net';
 import { Router } from '@angular/router';
 import { RitualService } from '../../services/ritual.service';
+import { ConfirmationModalComponent } from '../confirmation-modal/confirmation-modal.component';
+import { take } from 'rxjs';
 
 @Component({
   selector: 'app-menu',
@@ -16,6 +18,8 @@ export class MenuComponent {
   protected dialogRef = inject(DialogRef);
   protected router = inject(Router);
   protected ritualService = inject(RitualService);
+  protected dialog = inject(Dialog);
+  protected translateService = inject(TranslateService);
 
   protected version = signal('1.0.0');
 
@@ -28,11 +32,21 @@ export class MenuComponent {
     this.close();
   }
 
-  protected async deleteUserAccount(): Promise<void> {
-    const result = await this.ritualService.deleteUserAccount();
-    if (result) {
-      this.router.navigateByUrl('');
-      this.close();
-    }
+  protected deleteUserAccount(): void {
+    const modal = this.dialog.open(ConfirmationModalComponent, {
+      data: {
+        title: this.translateService.instant('confirmation-modal.delete-user.title'),
+        message: this.translateService.instant('confirmation-modal.delete-user.message'),
+      }
+    });
+    modal.closed.pipe(take(1)).subscribe( async (action) => {
+      if (action) {
+        const result = await this.ritualService.deleteUserAccount();
+        if (result) {
+          this.router.navigateByUrl('');
+          this.close();
+        }
+      }
+    });
   }
 }
